Декомпозиция проекта на модули
