+++ index.html
@@ -236,7 +236,8 @@
   </template>
 
 </body>
-<script src="functions.js"></script>
-<script src="main.js"></script>
+
+<script src="js/main.js" type="module"></script>
+<script src="js/functions.js"></script>
 
 </html>
