+++ js/constants.js
@@ -1,0 +1,40 @@
+export const PHOTO_DESCRIPTION_COUNT = 25;
+export const PHOTOS_MIN = 1;
+export const PHOTOS_MAX = 25;
+export const LIKES_MIN = 15;
+export const LIKES_MAX = 200;
+export const AVATAR_MIN = 1;
+export const AVATAR_MAX = 6;
+export const COMMENTS_MIN = 0;
+export const COMMENTS_MAX = 30;
+
+export const NAMES = [
+  'Настя_yellow',
+  'Kate',
+  'Иван',
+  'Маша',
+  'Olly_Sunshine',
+  'Женя',
+  'Олег',
+  'Кирилл Ivanov',
+  'Alex'
+];
+
+export const DESCRIPTIONS = [
+  'Прекрасный день!',
+  'Новая фотография для альбома.',
+  'Делюсь своими впечатлениями.',
+  'Захватывающее место!',
+  'Это был незабываемый момент.',
+  'Доброе утро, Мир!',
+  'Всем хорошего дня и прекрасного настроения.',
+];
+
+export const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
